+++ src/d2s/stats/mana.rs
@@ -19,8 +19,8 @@ impl Mana {
     pub fn build(character: &Character) -> Mana {
         Mana::new()
             .with_mana_current_id(8)
-            .with_mana_current_value(character.mana() * 256)
+            .with_mana_current_value((character.mana() * 256.0) as u32)
             .with_mana_max_id(9)
-            .with_mana_max_value(character.mana() * 256)
+            .with_mana_max_value((character.mana() * 256.0) as u32)
     }
 }
